+++ src/all.ts
@@ -1,4 +1,5 @@
 // Components
+export { CpsAccordion } from './components/accordion.js';
 export { CpsAvatar } from './components/avatar.js';
 export { CpsBackground } from './components/background.js';
 export { CpsBadge } from './components/badge.js';